requirements.txt
Django==1.6.5
https://bitbucket.org/andrewgodwin/south/get/e2c9102ee033.zip#egg=South
factory-boy
Unipath
gunicorn
django-braces
django-spurl
django-extensions==1.3.6
short-url==1.2.1
django-model-utils==2.0.3
dj-database-url==0.3.0
django-bootstrap-form
django-skd-tools
django-bootstrap3
django-separated
